requirements/dev.txt
-r base.txt
ipython==9.4.0
ipdb==0.13.13
django-debug-toolbar==6.0.0
django-extensions==4.1
black==25.1.0
django-stubs==1.2.0
